sdk docs: update RICH_TEXT and MULTIPLE_LOOKUP_VALUES in FieldType enum
Summary:
Now that RICH_TEXT is GAed, we can expose this in the FieldType enum.

Also added documentation for lookup fields, which was previously marked as "UNSTABLE" (I don't think that's the case anymore?)

I'll cut a hyperbase diff to update the actual docs once this is landed.

Test Plan: No qa needed.

Reviewers: somehats, kasrak, minicat

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D16504