sdk: create/update field
Summary:
Depends on D16227 and can only be landed once it's deployed to production.

Pretty similar to other mutations. The only weird thing is that `field.options` returns `null` instead of `undefined`, even though it's `undefined` in the underlying value: so I've done some manual translation to change it back to `undefined` when we pass it back to hyperbase (so field options can roundtrip safely).

Test Plan: Tested in test block (see D16227 for vids)

Reviewers: somehats, billylittlefield

Reviewed By: somehats

Subscribers: kasrak

Differential Revision: https://airtable.phacility.com/D16228