record_card: fix handling for onClick
Summary:
Noticed this from https://community.airtable.com/t/recordcard-override-onclick/28580

There are 2 slight bugs with `RecordCard` click handling. The expected behavior for `onClick` here is:
- `null` is a no-op
- `undefined` performs default behavior (expand record)
- any function provided will override the default behavior

Any other falsey values (false, "") should be treated like `null` IMO, and no-op.

The `hasOnClick` prop functions correctly when the RecordCard is wrapped in a Tooltip, but it doesn't work when the RecordCard is just on its own. In this scenario, it'll be undefined (assuming the user hasn't discovered this hidden prop). If `onClick` is overridden, we'll still end up doing the default behavior because `hasOnClick` is falsey. The fix is to explicitly check if the value is `false`, which will get set correctly by Tooltip. This way, it works both when wrapped and not.

Also fixes a similar bug in `RecordCardList`, where we were converting all falsey values into `undefined`, but we actually sometimes want to pass `null`.

Test Plan:
- Create a RecordCard, test behavior for `onClick` in 3 scenarios: `undefined`, `null`, and some custom function
- Wrap the RecordCard in a tooltip, and try the same thing

- Create a RecordCardList, test the same thing for `onRecordClick`.

All scenarios should behave consistently.

Reviewers: s2tephen, jay-ransijn

Reviewed By: s2tephen

Differential Revision: https://airtable.phacility.com/D16650