sdk: add @example back to metadata writes doc examples and regenerate docs
Summary:
Ran into an issue where the examples were rendering as part of the last `@param`
Somewhat luckily I'd missed removing one `@example` and that was the only one that worked correctly, so adding it back in for everything else.

On D16750 it was mentioned `@example` is optional now but seems like we need to fix the param case?

Also fixes a bug where an external link had the wrong syntax (oops)

Before:
{F174677}
After:
{F174678}

Test Plan: Regenerated the docs and tested against hyperbase

Reviewers: billylittlefield

Differential Revision: https://airtable.phacility.com/D17140