sdk: allow using UI library & unstable_private_utils in non-blocks contexts
Summary:
The primary change here is making SDK initialization and the creation of the aggregators library
lazy. This means that code-paths that import `get_sdk` don't automatically import every single
dependency of the SDK, and that aggregators don't try and read from airtable_interface until you
first use them.

The motivation here is to allow parts of the scripting block to run in hyperbase, outside of a
blocks-specific context. In general though, this will make a lot of things easier to run outside of
blocks contexts (e.g. in react-storybook and in tests).

I added a bunch of tests to make sure that the `ui` and `unstable_private_utils` entrypoints can be
imported without airtableInterface/the sdk already running, and to check that each of the UI
components used by the scripting block will render without airtableInterface/the rest of the SDK

Test Plan: Tested with a test block & the scripting block.

Reviewers: minicat, billylittlefield

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D17051