sdk: Relate FormField descs to input elems in a11y tree
Summary:
The `FormField` component supports the specification of input element
descriptions via the `description` prop. Extend the supported input elements to
detect such descriptions and to render with an appropriate value for the
`aria-describedby` property so that assistive technologies can detect this
relationship. Add tests for existing functionality to demonstrate that the
change does not introduce regressions.

Test Plan: automated tests included in patch

Reviewers: jay-ransijn

Reviewed By: jay-ransijn

Differential Revision: https://airtable.phacility.com/D17594