cli: output build errors on 'block release'
Summary:
Why
===
This is just an incremental improvement to having better console outputs on errors. In an ideal world, we'd have some "verbose" option that will output additional detaiils.

As a first step, this diff outputs to the terminal certain build-related errors on `block release`. These messages are always useful regardless of "verbose" mode or not, so we'll always output for now.

This is from feedback we've received a few times in the past couple months:
* https://h.slack.com/archives/CQJNQPEBB/p1587002388326900 (related [[ https://airtable.com/tblJvlqwVOgRLmgdT/viwGhVu5xW0h3vdiY/rec0ydHApJI75747U | escalation ticket ]])
* https://h.slack.com/archives/CSLFG7YJX/p1588008032026000 (related [[ https://airtable.com/tblJvlqwVOgRLmgdT/viwGhVu5xW0h3vdiY/rec9caTiBg6FYE9kX | escalation ticket ]])

What
====
* Output the error messages from `BlockBuilder` on `block release`.
* add some formatting to the error output

screenshots:
* Before {F182753}
* After {F182754}

Test Plan: No QA needed

Reviewers: minicat

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D18013