sdk: move RecordActionData to types & export
Summary:
Depends on D18353

I wanted to use it in the scripting block, but it's not currently exported.

It's a little sketchy to export it in types whilst the other stuff is still in `private_utils` but I think it's fine for this to show up in autocomplete for a little bit.

Test Plan: Tested that I could import it in scripting block

Reviewers: somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D18468