sdk: document record action APIs
Summary:
Depends on D18568

Add some examples and tidy up the documentation. Please review the documentation in full, not just the changed parts (I've been adding to it as the APIs are developed, but it'd be good to give it a proper review now).

They are still `@hidden`.

I'm hoping to include this in this week's SDK release so that we can link Formstack/Pexels these in the SDK repo. If it's not ready by then, I'll mirror to a doc we can share with them (I started a doc for the other button field related stuff to share with them: https://airtable.quip.com/pIbgAex7t6lB/Record-action-APIs-for-partners)

Test Plan: I'll test the examples in a block before landing

Reviewers: somehats, s2tephen, kasrak

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D18570