cli: Avoid false positive in test helper function
Summary:
The original implementation of `assertThrowsAsync` would not report an
error if the provided function thew an error synchronously. This
behavior actually described a failure state because consumers of
asynchronous APIs generally expect errors to be reported asynchronously.

Update the helper function to report a failure if the provided function
synchronously throws an error.

Test Plan: This change concerns the testing infrastructure itself.

Reviewers: richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D18706