Add script to fetch and apply patches from GH PRs
Summary:
Recently, an independent developer submitted a patch to one of the
public repositories which are created from the Block examples in this
repository. This script automates the process of retrieving such patches
and applying them to this repository. It does not involve publishing
such changes because the project already a procedure for creating new
releases.

Test Plan: No QA needed.

Reviewers: kasrak, richsinn, billylittlefield

Reviewed By: richsinn

Subscribers: somehats

Differential Revision: https://airtable.phacility.com/D18895