ui kit: explicitly use 'border-box' box-sizing on ChoiceToken inner div
Summary: Addresses issue found in https://airtable.phacility.com/D19459, where the `box-sizing` for the inner div of a `ChoiceToken` was inheriting `box-sizing` of `content-box` instead of `border-box`, causing the component to be rendered incorrectly. This explicitly adds the correct `box-sizing` to the inner div to prevent other overwrites.

Test Plan: No qa needed

Reviewers: jay-ransijn, s2tephen

Reviewed By: s2tephen

Differential Revision: https://airtable.phacility.com/D19545