Video Snippets Block: work around browser bug
Summary:
Prior to this patch, a bug in the Chromium browser causes the Block to
crash when transitioning between "Full Screen" mode in Chromium, Google
Chrome, and (presumably) Microsoft Edge. In these situations, the
browser was firing a global "error" event, and the Blocks framework
interpreted this as an unrecoverable problem.

Modify layout of the browser's native "video" element UI widget in order
to prevent this error from being reported.

Test Plan:
1. Run the Video Snippets Block in Chrome version 83 or later
2. Expand the browser window to a width of at least 1500 pixels
3. Select a video
4. Enter "Full Screen" mode
5. Exit "Full Screen" mode

No error should be reported

Reviewers: kasrak, richsinn

Reviewed By: richsinn

Differential Revision: https://airtable.phacility.com/D19631