Revert "cli: Remove dep. on `git` executable using NodeGit"
Summary:
This reverts commit bc39e6f5b9dbac6861675dafb62f1cacda8a85a9. (D19398)

This is only a partial revert. It keeps the following changes from D19398:
* The TypeScript version from the root `package.json` because version 3.9.5 is incompatible with some SDK files. We should fix/address this in a separate diff; it is outside the scope of these CLI changes. See [[ https://airtable.phacility.com/D19398#463640 | comment ]] in D19398.
* Some cosmetic method naming changes in test file (add `Async` suffix to method name)

We are reverting because the `nodegit` package requires the macOS command line tools to be installed if using the CLI tool from macOS (because it needs to build nodegit with gyp).

We figured it's better to maintain the `git` dependency for now, instead of having the macOS command line tools dependency for macOS users.

TODO: We are investigating a way to remove the `git` dependency via using the GitHub APIs directly.

Test Plan: No QA needed

Reviewers: kasrak, minicat, billylittlefield, jugglinmike

Reviewed By: minicat

Subscribers: kasrak

Differential Revision: https://airtable.phacility.com/D19865