Add new preview types to Url Preview Block
Summary:
Users requested many different url preview types in a form embedded in the block itself.

I went through every type of requested URL to determine how viable it was to add a url preview to this block.

https://airtable.com/tblncR9PaL36GtcEr/viwK2SQ5SvrDRxesU?blocks=bipSgzxRbpEGfjaZK

Most requests were not viable i.e. the website provides no easily viewable preview/embed link.

Some were straightforward to implement via the current meth of matching the raw url and converting it to a preview url and placing it in an iframe.

I implemented for Google Docs, Google Sheets, Google, Google Drive, Google Slides and Loom.

For the Google previews, I used Google's preview view. Google additionally provides an embedable link that looks even nicer than the preview, but for the embeded link to work the owner of the document must explicately "publish" the
document, so I think it is preferable to show the preview.

In addition to the above previews were added to urls that go to arbitary images. The only way to really know a url refers to an image is to load the image and see if it succeeds or not. This is a async operation so the RecordPreview had
to be refactored so previewUrl was state that could be set asyncly.

There were many other Url Previews that were requested that were technically viable but where not straightforward to add, so I did not add them. None of these could work via taking a regex of the raw url generating another preview url and dropping it in an iFrame.

There are two main categories:
1. "Yes, its complicated":* Most of these require custom javascript to be loaded in the iframe. Some require API keys. I found and linked to the developer documentation for each of these in the Notes section.

2. "Yes, with oembed": All of these require more than a url to be included in an iframe, but instead of being custom for each website they follow a standard called [oembed](https://oembed.com/). That means if we implemented oembed we could get support not just for the 11 requested of url preview types but all 250+ providers of the oauth standard. It would be a bit of work to add oembed support to this app, but I think it would be worthwhile because not every website would need to be individually implemented. If you think this work is worthwhile I think we should follow up with it, probably in another PR.

I added new example urls that cannot be tested here:
https://airtable.com/tblAFhjuXNkjVAqN4/viwqgGwiSeV9d4TJ4?blocks=bipSgzxRbpEGfjaZK

I just tested the block locally. I'm not sure if I should be releasing it. I assume it should be merged first.

Test Plan:
QA the Preview Block.

Here is the QA plan that was orignally created for the block I added and highlighted the new URL types that are supported: https://docs.google.com/document/d/1nob0gbUyk82xxiPcLE1aRGngf2bGHA5rEKnGH3BPhOg/edit

Reviewers: richsinn, rmeritz

Differential Revision: https://airtable.phacility.com/D20194