sdk: make all AirtableInterface usage lazy
Summary:
This is a (very late) follow-up to D17051. It replaces all our usage of airtableInterface with a call to `getAirtableInterface`, which will initialize AirtableInterface if it hasn't been already. The goal here, like in D17051, is to make it easier to import & use parts of the SDK that aren't specific to the blocks outside of the blocks environment.

This is part of a longer-term goal to get rid of `getSdk()` entirely...

Generated using these two codemod transforms:
- https://github.com/Hyperbase/codemods/blob/d2b6743e3bec2cf198dde9b051fb9733f876bb05/src/transforms/airtableInterfaceToCall.ts
- https://github.com/Hyperbase/codemods/blob/d2b6743e3bec2cf198dde9b051fb9733f876bb05/src/transforms/noSdkAirtableInterface.ts

Test Plan: QA can skip.

Reviewers: minicat

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D17372