Extend mock to test SDK more fully
Summary:
Extend the functionality of the mock `AirtableInterface` so more of the
interaction between a Block and a Base can be verified in automation.
Introduce new tests to demonstrate testing the SDK's response to
messages from the Base. Extend some existing tests to demonstrate
testing the SDK's response to developer-facing methods.

Test Plan: n/a (this changeset concerns the tests themselves)

Reviewers: richsinn, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D20431