sdk: blocks to apps in block_wrapper
Summary:
Updates the text in the block_wrapper "Please make this block bigger or fullscreen" to use app/block depending on feature flag.

Also does a very light update of the README.md to not explicitly reference "custom blocks". Happy to remove this change if we consider it out of scope.

Test Plan:
1. Create a block using this updated version of the SDK
2. Add a <ViewportConstraint />
3. Verify the copy changes according to the feature flag

Reviewers: #team-platform-ecosystem, s2tephen

Reviewed By: #team-platform-ecosystem, s2tephen

Differential Revision: https://airtable.phacility.com/D21951