(2/3) Update SDK to allow detecting whether enterprise has enabled block development
Summary:
Part 2 of 3

Part 1: D22009
Part 3: D22024

The SDK needs to expose to blocks whether or not block development should be restricted (due to enterprise setting). This is in particular needed for the scripting block.

Test Plan: When building an app with this version of the sdk, the `isBlockDevelopmentRestrictionEnabled` method should be available on unstable_private_utils. This is best tested end-to-end via the test plan in  D22024

Reviewers: benrazon, minicat

Reviewed By: benrazon, minicat

Subscribers: somehats

Differential Revision: https://airtable.phacility.com/D22011