sdk: fix select non-unique key bug with duplicate values
Summary:
For this issue: https://airtable.com/tbl3f9cb8b2f15ab8/viw7P84NHWqHTImvH/recxzsKe17GiY3lip?blocks=hide

Previously, we'd use a wrapped form of the `value` as the `key` for each select option - this is problematic when the `value` is non unique. This manifested as a bug in script settings, where, if you had multiple options with the same value (for example, you copy and paste an option) and then edit them, the old ones aren't properly updated and it looks like there are duplicates of the options. There's also a console warning about this behaviour due to the duplicate keys.

My assumption is that this usually isn't a problem as you'd have unique values, but I think this could come up somewhat frequently with script settings in that above example scenario.

AFAICT this is a long standing issue, dating back to the original implementation: https://github.com/Hyperbase/hyperbase/commit/0c52816cc8df069158be41ef680e6aea1eb99044 I'm not sure if there's a specific reason to use `valueJson` but I think using `index` here is safe.

`SelectButtons` has the same issue - also fixed it here.

Test Plan:
I tested against the script block and this fixes the bug.

I also checked Storybook for both `Select` and `SelectButtons`.

Reviewers: s2tephen, jay-ransijn, billylittlefield

Reviewed By: s2tephen, jay-ransijn, billylittlefield

Differential Revision: https://airtable.phacility.com/D22835