sdk: Improve test coverage for Base model
Summary:
Improve consistency of mock usage

One test for the Base model mocked an instance of the Sdk and of the
Mutations model. Replace this with a lower-level mock of the
AirtableInterface in order to reduce the amount of fake code under test
and to improve consistency with the mocking strategy used by other
tests.

Configure the code coverage tool to ignore one branch related to an
internal optimization.

Test Plan: n/a (this changeset concerns the tests themselves)

Reviewers: richsinn, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D21565