sdk: Enforce test coverage for models
Summary:
Configure the Jest test runner to analyze test coverage for the SDK's
models and to require complete coverage. Tolerate the existing gaps in
coverage with in-line directives. These directives may be removed
incrementally as coverage is improved; in the mean time, this change
will help avoid regression in existing files and promote thorough
testing of new files.

Test Plan: n/a (this changeset concerns the tests themselves)

Reviewers: richsinn, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D23040