sdk: Remove source of unhandled Promise rejection
Summary:
Prior to this commit, the `loadDataAsync` method of the `RecordStore`
model invoked `.then` and discarded the resulting promise.

This risked producing unhandled rejection errors. Firefox, Chromium, and
Node.js all issue a warning message in this event, and this may be
distracting for consumers (especially because they cannot silence it).
Additionally, Node.js version 12 warns that a future version of the
platform will force the process to terminate.

The effect of the unhandled rejection was otherwise unobservable because
the call site used the internal `fireAndForgetPromise` to convert
Promise rejections to uncaught exceptions. This obviates the need for an
automated test. Although including additional fallible operations in the
Promise chain increases the number of ways the Promise may be rejected,
the use of `fireAndForgetPromise` makes this change in semantics
unobservable.

Test Plan:
n/a (this patch is intended to improve developer ergonomics and avert
failure in future versions of the testing platform)

Reviewers: richsinn, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D23155