sdk: Add tests for LinkedRecordsQueryResult model
Summary:
Description:
Introduce a mechanism for temporarily disabling object pooling from the
test environment. Utilize that to more completely test the model.

Extend the AirtableInterface mock with APIs which are exercised by the
model but which had not yet been implemented.

Configure the test coverage enforcement tool to tolerate specific gaps
in coverage; include in-line comments to document the rationale for each
disablement.

Test Plan: n/a (this changeset concerns the tests themselves)

Reviewers: richsinn, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D23760