sdk: Refactor ObjectPool
Summary:
Simplify the requirements imposed on objects that may be pooled. Express
this as a formal TypeScript interface named "Poolable," and update the
poolable objects accordingly.

Previously, each poolable object defined a set of parameters that would
be provided when an object was requested. These determined if the
request could be satisfied by an existing pool member or if a new
instance would need to be created. The parameters differed only slightly
from the parameters used in object construction. Simplify this algorithm
by creating a new instance in all cases and by using the "key" of the
newly-created object to determine whether it should be used or
discarded. (This essentially relocates the responsibilities of the
`_canObjectBeReusedForOptions` function to the pool key, so in removing
the former, the latter must be extended for some poolable objects.)

Previously, the classes of poolable objects implemented static methods
named  `__createOrReuseQueryResult` for creating instances. These
methods were called by the models which maintained the object pools, and
they ultimately called back out to those models. Simplify the control
flow and division of responsibility by relocating the logic for creating
poolable objects to the models which maintain the associated object
pools.

Test Plan:
the behavior of the ObjectPool is asserted indirectly through
the existing model tests and UI tests.

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24324