sdk: Reduce reliance on global state
Summary:
Move ObjectPool for RecordQueryResults to an instance property of the
Table class.

sdk: Reduce reliance on global state

Move ObjectPool for ViewMetadataQueryResults to an instance property of
the View class.

sdk: Reduce reliance on global state

Move ObjectPool for LinkedRecordsQueryResults to an instance property of
the Record class.

sdk: Remove recently-obviated test helpers

The structure of the ObjectPool was recently refactored to facilitate
testing. This makes the helper functions removed by this patch
(implemented to subvert global state) unnecessary.

With this patch applied, each record maintains a dedicated pool for
LinkedRecordQueryResults. There is no longer a risk of collision across
records, so the object keys do not need to include record identifiers. Simplify
the keys used to insert results into these pools.

Test Plan:
n/a (this refactoring is verified by previously-written
automated tests)

n/a (this refactoring is verified by previously-written
automated tests)

n/a (this refactoring is verified by previously-written
automated tests)

n/a (this patch concerns the tests themselves)

Reviewers: richsinn, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24148