CLI: include source maps when using block run
Summary:
Block run is only used during development, so having source maps is useful for debugging.
I believe there is no need to add a flag for this and we have a separate step that minifies
release builds so this won't cause us to leak source code for our blocks publicly.

Test Plan:
Cause a block to crash in development mode and see the better stack trace.
Cause a block to crash in release mode and see that the stack trace and sources
are still minified.

Reviewers: somehats, richsinn

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24355