sdk: Reconcile changes from latest patches
Summary:
Two commits [1] [2] which were recently merged independently have
unanticipated interdependencies.

One commit removed a testing API which was in use by the other. In
addition, that commit also removed the final reference to an accessor
property (`__sourceModelId`). Because the second commit enabled code
coverage enforcement for that file, the test suite did not flag the
newly-unused code path until the patches were combined.

Configure the test coverage tooling to tolerate the unused accessor
property. Remove the references to the obviated testing API.

[1] fedffa9fb9dce21599cc47ca9e9ef25c9c825830 - sdk: Refactor ObjectPool
[2] 604d9871e06c57df4f947a1306a606c2de2ec364 - sdk: Add tests for TableOrViewQueryResult model

Test Plan: n/a (this changeset concerns the tests themselves)

Reviewers: richsinn, somehats, benrazon

Reviewed By: richsinn, benrazon

Differential Revision: https://airtable.phacility.com/D24487