sdk: Remove faulty optimization from Base model
Summary:
Prior to this commit, the Base model circumvented lazy Table
instantiation when propagating events from AirtableInterface. The
associated inline code comment correctly explained that the absence of
the Table models proved that there could be no consumers of Table
events. This was taken as proof that the event triggering logic did not
need to run.

However, under these conditions, there could be consumers of Base events
(as demonstrated by existing unit tests). The "schema" Base event is
dependent on the propagation of Table events, so by neglecting to
propagate Table events (even in the absence of consumers), the
implementation failed to produce relevant Base events.

Propagate Table events regardless of the presence of consumers so that
consumers listening to the "schema" Base event are correctly notified.

Test Plan: automated by included unit tests

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24576