sdk: Fix bug in LinkedRecordsQueryResult
Summary:
Because they operated after model keys had been unwatched, the methods
named `_unwatchLinkedQueryCellValuesIfPossibleAfterUnwatch`
`unwatchLinkedQueryCellValuesInFieldIfPossibleAfterUnwatch` were
incapable of discriminating between two conditions:

- the removal of the final subscriber for a given key
- the superfluous removal of a non-existent subscriber from an unwatched
  key

The model is intended to clear some internal state in response to the
former condition. It is intended to throw an error in response to the
latter condition.

This distinction can only be made prior to unwatching a given key.

Refactor and rename the methods for use prior to unwatching and
enable/extend the previously-submitted tests designed to verify the
intended behavior.

Test Plan:
The correction concerns an internal optimization that is not
directly observable. The included automated tests verify that it
operates as expected.

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24519