sdk: Avoid crash on release of weakly-held object
Summary:
Prior to this patch, the `unregisterObjectForReuseWeak` was invoked in a
new call stack. When it threw an error under these conditions, the error
could not be caught and would ultimately cause the Airtable App to
crash.

A recent refactoring of the ObjectPool model removed all external
callers of this method, making it safe for this patch to remove the
method entirely. Replace with the internal version which performs the
same operation without throwing.

Test Plan: automated by included unit tests

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24597