check typescript in the context of a block in CI
Summary:
When we bundle and release blocks, we strip typescript implementations and just leave the type declarations behind. Type declarations marked as `@internal` also get stripped, which can sometimes lead to problems where we pass typechecks when developing locally, but fail once the SDK is packaged up and installed in an actual project.

This test also checks against the latest version of typescript. This is a little fragile, but is deliberate as generally i think we should aim to support the latest typescript. If that changes, we can update the tests.

Included in this diff is a fix to `LinkedRecordQueryResult` for an issue along these lines.

Test Plan: -

Reviewers: jugglinmike, minicat

Reviewed By: jugglinmike

Differential Revision: https://airtable.phacility.com/D24673