sdk: Correct regressions from recent commit
Summary:
A recent commit [1] fixed a bug relating to the way the "deleted" status
of models was being managed. The change had two unintended side-effects:

1. It caused the SDK to produce less helpful error messages in some
   conditions. Where it previously reported "RecordQueryResult's
   underlying view has been deleted" it instead reported
   "RecordQueryResult data is not loaded"
2. It made it impossible to load LinkedRecordsQueryResult models

Both of these changes caused test failures in the latest version of the
project's automated test suite. Neither were identified during the
review process because the patch was written against a prior revision of
the codebase which predated the relevant tests.

For similar reasons, the automated validation process failed to
recognize a newly-introduced unreachable branch.

Restructure the order of operations in `RecordQueryResult#recordIds` so
that it reports the most salient of the errors which explain failure.
Extend `LinkedRecordsQueryResult` to support the new checks for
deletion. Remove the unreachable branch.

[1] bfb494d4ac5434ea8d07fe0b22b2300eb487bd03

Test Plan: automated by included unit tests

Reviewers: minicat, somehats, MananNayak-at

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24756