sdk: Improve error msg for loading w/deleted src
Summary:
Previously, when the "source" model for a RecordQueryResult instance was
deleted, attempting to load the query result's data would produce an
error in terms of the underlying RecordStore. Although the description
was technically accurate, it was not particularly meaningful to
consumers of the SDK because the RecordStore model is not part of the
public API.

Insert a guard in the model loading logic so that a more helpful error
message (one written in terms of Views and Tables) is reported.

Test Plan: automated by included unit tests

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24880