sdk: Refactor references to SDK instance
Summary:
This patch is in service of an ongoing effort to enable third-party App
developers to write automated tests for their applications.

Prior to this patch, the runtime's single Sdk instance was accessible
via a module named `get_sdk`. This pattern subverted the project's
dependency tree which in turn complicated the automated tests. The
complication could be addressed with additional testing infrastructure,
and although this might be satisfactory for internal tests, it would be
less desirable in the context of tests for third-party Apps.

Explicitly manage references to the runtime's Sdk instance. For
JavaScript models, this means passing `sdk` references via constructor
parameters. For React Components, this means introducing a Context
component.

This changeset includes additional changes which are necessary to
preserve existing behavior:

- Remove codepaths which are no longer necessary.
- Deprecate exported bindings which necessitate cycles in the project's
  dependency graph.
- Create a new entry point (`ui/unstable_standalone_ui`) for the React
  Components which can be instantiated outside of an Airtable Apps
  context.

Add automated tests to verify previously-untested contracts and reduce
the likelihood of regression due to this patch or any subsequent patch.

Test Plan: automated by included tests

Reviewers: somehats, benrazon

Reviewed By: somehats

Subscribers: somehats

Differential Revision: https://airtable.phacility.com/D25028