sdk: Statically enforce inheritance structure
Summary:
Previously, a number of internal classes were not intended to be
instantiated and implemented some methods which had to be overridden by
children. The first expectation was not enforced in any way, and the
second was only enforced at runtime via explicit `throw` statements
(which needed to be ignored by the test coverage analysis tooling).

The TypeScript language offers a pattern for abstract classes and
abstract methods to express (and enforce) exactly this structure.
Re-implement the relevant classes and methods to be "abstract". Update
the classes' descendants to nominally conform to this contract in those
cases where they intentionally violate it.

Test Plan:
n/a (this change only effects code paths which were
previously unreachable)

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D24946