sdk: Prevent & disallow unhandled test rejections
Summary:
Prior to this patch, the tests for the Table model prompted the creation
of a number of Promise values which were rejected and unhandled (either
by neglecting to mock parts of AirtableInterface or by failing to handle
an explicitly-created rejected Promise). This caused the Node.js runtime
to print a large amount of warning messages to the screen, e.g.

>     DeprecationWarning: Unhandled promise rejections are deprecated.
>     In the future, promise rejections that are not handled will
>     terminate the Node.js process with a non-zero exit code.

The behavior described by this message represented a hazard for
upgrading--although passing in today's release of Node.js, the test
suite could be expected to fail in some future version.

In addition, these messages obscured information about the progress of
the test runner.

Correct the tests to avoid creating unhandled rejections. Update the
infrastructure to opt in to the future behavior so that future
contributions do not unknowingly introduce regressions along these
lines.

Test Plan: n/a (this change concerns the tests themselves)

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D25165