CLI: Add uploadSourceMapsToRollbar flag to release command
Summary:
This adds a hidden flag to the block release command that allows uploading
the block's source code to rollbar.
To use this the block must set bundleCdn in its remote.json file. The block
is responsible for setting up rollbar with the version set to it's git sha.
This is currently hard to do, but I'm planning on adding a constant that the cli
will inject to make this easier in a follow up commit.
https://rollbar.com/Airtable/Airtable-Blocks/items/55247/ is an example rollbar
with this setup. There are some ugly relative paths do to how we implement isolated
builds, but I think this is still a big improvement and we can work on fixing this if
it causes issues with rollbar coalescing errors.

Test Plan:
I tested manually in development by releasing and publishing a block with this cli
with the flag and checking that the rollbar included source maps.

Reviewers: somehats, billylittlefield

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D25277