Cli: suppress git output
Summary:
In order to set up source maps I made the cli include the git hash in a window
variable. However, not everyone has a repo set up for their block and even
if they do, I don't think there's any reason to display the git hash just because
we get it via shelling out.
I was tempted to do this for the npm output as well, but I'm concerned that it
is sometimes really helpful if something goes wrong.

Test Plan: I ran block init and saw that there's no longer a git error printed.

Reviewers: billylittlefield, SeanKeenan-at

Reviewed By: billylittlefield

Differential Revision: https://airtable.phacility.com/D25619