Add performBackendFetch method to sdk
Summary:
https://docs.google.com/document/d/1v1u1ffNf8HME1wddigT0cgbdwN9QvICpmyWJ5tGjMn0/edit?ts=5f2022be#
This adds a method to the sdk so that you can use the backend fetch implementation and matches up with the
addition to airtable interface in https://airtable.phacility.com/D25167
I left it as a hidden method so that we can dogfood it with the scripting block first (and maybe some other first party blocks), but I expect that using this interface will be what we end up exposing to third party developers. It seems much simpler than intercepting fetch calls with mode set to proxy, gives us more flexibility if we need to add features, and doesn't seem any worse in terms of developer experience.

Test Plan: No manual QA needed.

Reviewers: somehats

Reviewed By: somehats

Subscribers: yoel-airtable, djmally

Differential Revision: https://airtable.phacility.com/D21007